Updated Java call to use proper classpath separator on Windows.
On Linux and MacOS, classpath entries are separated by colons. On
Windows, the colon is a special character to specify root drives.
Because of this, on Windows, the classpath separator is a semicolon.

This means that the MIST load must be 'plugins/MIST_.jar;jars/*'
ONLY on Windows, but can be 'plugins/MIST_.jar:jars/*' everywhere
else.

This is made 100x more horrifying by much of the Java documentation
assuming Windows as the default.