Made CellProfiler calls OS-independent and removed spurious prints.
Two changes were made: first, the CellProfiler call was changed
to be fully quoted ({cp_app:q}). This is to ensure that the correct
binary can be found, even when the path to CellProfiler contains
spaces.

The second change is in the log redirection. Previously, log redirection
had used "&> {log}", however, this is a bash-ism and only works on Bash.
Now, we use the verbose "1> {log:q} 2>&1", which should work on
every major shell on every major OS nowadays.